重新将通知监听服务合并至主进程，因为独立进程的运行状态有问题；去除通知使用权是否授予的监听器

+++ app/src/main/java/com/manager/assistant/ManagerAssistant.java
@@ -2,6 +2,7 @@ package com.manager.assistant;
 
 import android.app.Application;
 import android.net.Uri;
+import android.os.Build;
 import android.util.Log;
 
 import androidx.work.WorkInfo;
@@ -26,41 +27,43 @@ public class ManagerAssistant extends Application {
     public void onCreate() {
         super.onCreate();
 
-        //初始化动态配色
-        if (AppSettingsPreference.getDynamicColorStat(this)) {
-            DynamicColorsOptions options = new DynamicColorsOptions.Builder()
-                    .setThemeOverlay(R.style.Theme_ManagerAssistant_Dynamic)
-                    .build();
-            DynamicColors.applyToActivitiesIfAvailable(this, options);
-        }
-
-        //安排自动备份任务
-        if (AutoBackupPreference.getSwitchStat(this)) {
-            int frequency_index = AutoBackupPreference.getBackupFrequency(this);
-            long intervalMillis = AutoBackupHelper.BackupFrequency.values()[frequency_index].getIntervalMillis();
-            BackupScheduler.schedulePeriodicBackup(this, intervalMillis);
-
-            //打印任务状态日志
-            WorkInfo info;
-            try {
-                info = WorkManager.getInstance(this).
-                        getWorkInfosForUniqueWork(BackupScheduler.BACKUP_WORK_NAME).
-                        get().get(0);
-            } catch (ExecutionException | InterruptedException e) {
-                throw new RuntimeException(e);
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P && getProcessName().equals(getPackageName())) {
+            //初始化动态配色
+            if (AppSettingsPreference.getDynamicColorStat(this)) {
+                DynamicColorsOptions options = new DynamicColorsOptions.Builder()
+                        .setThemeOverlay(R.style.Theme_ManagerAssistant_Dynamic)
+                        .build();
+                DynamicColors.applyToActivitiesIfAvailable(this, options);
             }
-            Log.d(LogTags.WORK_STATS.getV(), "State: " + info.getState());
-        }
 
-        //启动时检测是否有需要删除的按转包
-        String apkUri = VersionPreference.getApkUri(this);
-        if (!apkUri.isEmpty()) {
-            File apkFile = new File(Objects.requireNonNull(Uri.parse(apkUri).getPath()));
-            if (apkFile.exists() && apkFile.delete()) {
-                Log.d(LogTags.APPLICATION.getV(), String.format(Locale.getDefault(), "成功删除“%s”", apkFile.getName()));
-                VersionPreference.setApkUri(this, "");
-            } else {
-                Log.w(LogTags.APPLICATION.getV(), String.format(Locale.getDefault(), "“%s”删除失败", apkFile.getName()));
+            //安排自动备份任务
+            if (AutoBackupPreference.getSwitchStat(this)) {
+                int frequency_index = AutoBackupPreference.getBackupFrequency(this);
+                long intervalMillis = AutoBackupHelper.BackupFrequency.values()[frequency_index].getIntervalMillis();
+                BackupScheduler.schedulePeriodicBackup(this, intervalMillis);
+
+                //打印任务状态日志
+                WorkInfo info;
+                try {
+                    info = WorkManager.getInstance(this).
+                            getWorkInfosForUniqueWork(BackupScheduler.BACKUP_WORK_NAME).
+                            get().get(0);
+                } catch (ExecutionException | InterruptedException e) {
+                    throw new RuntimeException(e);
+                }
+                Log.d(LogTags.WORK_STATS.getV(), "State: " + info.getState());
+            }
+
+            //启动时检测是否有需要删除的按转包
+            String apkUri = VersionPreference.getApkUri(this);
+            if (!apkUri.isEmpty()) {
+                File apkFile = new File(Objects.requireNonNull(Uri.parse(apkUri).getPath()));
+                if (apkFile.exists() && apkFile.delete()) {
+                    Log.d(LogTags.APPLICATION.getV(), String.format(Locale.getDefault(), "成功删除“%s”", apkFile.getName()));
+                    VersionPreference.setApkUri(this, "");
+                } else {
+                    Log.w(LogTags.APPLICATION.getV(), String.format(Locale.getDefault(), "“%s”删除失败", apkFile.getName()));
+                }
             }
         }
     }
